Only publish `DistributedEventSent/DistributedEventReceived ` events in unit test.
